+++ kwcoco/util/__init__.py
@@ -4,11 +4,13 @@ mkinit ~/code/kwcoco/kwcoco/util/__init__.py -w
 
 from kwcoco.util import dict_like
 from kwcoco.util import jsonschema_elements
+from kwcoco.util import lazy_frame_backends
 from kwcoco.util import util_archive
 from kwcoco.util import util_delayed_poc
 from kwcoco.util import util_futures
 from kwcoco.util import util_json
 from kwcoco.util import util_monkey
+from kwcoco.util import util_resources
 from kwcoco.util import util_sklearn
 from kwcoco.util import util_truncate
 
@@ -19,19 +21,21 @@ from kwcoco.util.jsonschema_elements import (ALLOF, ANY, ANYOF, ARRAY, BOOLEAN,
                                              OBJECT, ONEOF, QuantifierElements,
                                              STRING, ScalarElements,
                                              SchemaElements, elem,)
+from kwcoco.util.lazy_frame_backends import (LazyGDalFrameFile,
+                                             LazyRasterIOFrameFile,
+                                             LazySpectralFrameFile, )
 from kwcoco.util.util_archive import (Archive, unarchive_file,)
 from kwcoco.util.util_delayed_poc import (DelayedChannelConcat, DelayedCrop,
                                           DelayedFrameConcat, DelayedIdentity,
                                           DelayedImageOperation, DelayedLoad,
                                           DelayedNans, DelayedVideoOperation,
-                                          DelayedVisionOperation, DelayedWarp,
-                                          LazyGDalFrameFile, have_gdal,
-                                          profile, validate_nonzero_data,)
+                                          DelayedVisionOperation, DelayedWarp)
 from kwcoco.util.util_futures import (Executor, JobPool,)
 from kwcoco.util.util_json import (IndexableWalker, ensure_json_serializable,
                                    find_json_unserializable,
                                    indexable_allclose,)
 from kwcoco.util.util_monkey import (SupressPrint,)
+from kwcoco.util.util_resources import (coerce_num_workers,)
 from kwcoco.util.util_sklearn import (StratifiedGroupKFold,)
 from kwcoco.util.util_truncate import (smart_truncate,)
 
@@ -41,11 +45,13 @@ __all__ = ['ALLOF', 'ANY', 'ANYOF', 'ARRAY', 'Archive', 'BOOLEAN',
            'DelayedLoad', 'DelayedNans', 'DelayedVideoOperation',
            'DelayedVisionOperation', 'DelayedWarp', 'DictLike', 'Element',
            'Executor', 'INTEGER', 'IndexableWalker', 'JobPool',
-           'LazyGDalFrameFile', 'NOT', 'NULL', 'NUMBER', 'OBJECT', 'ONEOF',
+           'LazyGDalFrameFile', 'LazyRasterIOFrameFile',
+           'LazySpectralFrameFile', 'NOT', 'NULL', 'NUMBER', 'OBJECT', 'ONEOF',
            'QuantifierElements', 'STRING', 'ScalarElements', 'SchemaElements',
-           'StratifiedGroupKFold', 'SupressPrint', 'dict_like', 'elem',
-           'ensure_json_serializable', 'find_json_unserializable', 'have_gdal',
-           'indexable_allclose', 'jsonschema_elements', 'profile',
+           'StratifiedGroupKFold', 'SupressPrint', 'coerce_num_workers',
+           'dict_like', 'elem', 'ensure_json_serializable',
+           'find_json_unserializable', 'indexable_allclose',
+           'jsonschema_elements', 'lazy_frame_backends',
            'smart_truncate', 'unarchive_file', 'util_archive',
            'util_delayed_poc', 'util_futures', 'util_json', 'util_monkey',
-           'util_sklearn', 'util_truncate', 'validate_nonzero_data']
+           'util_resources', 'util_sklearn', 'util_truncate']